consolidating stories
